+++ clients/android/app/src/main/java/com/pichat/android/ui/viewmodel/ChatViewModel.kt
@@ -214,6 +214,6 @@ class ChatViewModel(application: Application) : AndroidViewModel(application) {
 
     override fun onCleared() {
         super.onCleared()
-        repository.disconnect()
+        repository.close()
     }
 }